+++ components/navigation/toolingItems.tsx
@@ -6,7 +6,15 @@ import IconHub from '../icons/Hub';
 import IconModelina from '../icons/Modelina';
 import IconParser from '../icons/Parser';
 
-const toolingItems = [
+interface ToolingItem {
+  href: string;
+  icon: React.ComponentType<any>;
+  title: string;
+  description: string;
+  beta?: boolean;
+}
+
+const toolingItems: ToolingItem[] = [
   { href: 'https://studio.asyncapi.com/', icon: IconHub, title: 'Studio', description: 'Visually design your AsyncAPI files and event-driven architecture.', beta: true },
   { href: '/tools/generator', icon: IconGenerator, title: 'Generator', description: 'Use your AsyncAPI files to generate documentation, code, anything!' },
   { href: '/tools/cli', icon: TerminalIcon, title: 'CLI', description: 'Interact with AsyncAPI from the comfort of your CLI.' },
