--- types/form/Select.ts
@@ -1,11 +1,0 @@
-export interface Option {
-    value: string;
-    text: string;
-}
-
-export interface SelectProps {
-    className?: string;
-    onChange?: (selected: string) => void;
-    options: Option[];
-    selected?: string;
-}
